requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.26.1
Pillow
huggingface_hub
einops
safetensors
evo
open3d
matplotlib
scipy
opencv-python
scikit-image
tqdm
trimesh
pycolmap==3.10.0
pyceres
